+++ site-cookbooks/jenkins/recipes/job.rb
@@ -7,13 +7,28 @@
 # All rights reserved - Do Not Redistribute
 #
 
-node[:jenkins][:job][:polling].each do |job_name|
+node[:jenkins][:job][:check_pull_request].each do |job_name|
   template 'tmp/config.xml' do
-    source 'jobs/polling.xml.erb'
+    source 'jobs/check_pull_request.xml.erb'
     owner 'root'
     group 'root'
     mode '0755'
-    variables app_info: node[job_name]
+    variables app_info: node[job_name.split('-').first],
+              token: Chef::EncryptedDataBagItem.load('github', 'token')['public_repo']
+  end
+
+  upsert_job do
+    job_name job_name
+    file_name 'tmp/config.xml'
+  end
+end
+
+node[:jenkins][:job][:denebola].each do |job_name|
+  template 'tmp/config.xml' do
+    source "jobs/#{job_name}.xml.erb"
+    owner 'root'
+    group 'root'
+    mode '0755'
   end
 
   upsert_job do
@@ -37,19 +52,47 @@ node[:jenkins][:job][:deploy].each do |job_name|
   end
 end
 
-node[:jenkins][:job][:check_pull_request].each do |job_name|
-  template 'tmp/check_pull_request.xml' do
-    source 'jobs/check_pull_request.xml.erb'
+node[:jenkins][:job][:polling].each do |job_name|
+  template 'tmp/config.xml' do
+    source 'jobs/polling.xml.erb'
     owner 'root'
     group 'root'
     mode '0755'
-    variables app_info: node[job_name.split('-').first],
-              token: Chef::EncryptedDataBagItem.load('github', 'token')['public_repo']
+    variables app_info: node[job_name]
   end
 
   upsert_job do
     job_name job_name
-    file_name 'tmp/check_pull_request.xml'
+    file_name 'tmp/config.xml'
+  end
+end
+
+node[:jenkins][:job][:update_gems].each do |job_name|
+  template 'tmp/config.xml' do
+    source 'jobs/update_gems.xml.erb'
+    owner 'root'
+    group 'root'
+    mode '0755'
+    variables app_info: node[job_name.split('-').first]
+  end
+
+  upsert_job do
+    job_name job_name
+    file_name 'tmp/config.xml'
+  end
+end
+
+node[:jenkins][:job][:zosma].each do |job_name|
+  template 'tmp/config.xml' do
+    source "jobs/#{job_name}.xml.erb"
+    owner 'root'
+    group 'root'
+    mode '0755'
+  end
+
+  upsert_job do
+    job_name job_name
+    file_name 'tmp/config.xml'
   end
 end